Separated gulfstream from ssh, added comments to plotting section of ssh_eval.py
